requirements.txt
carto==1.0.1
click==6.7
PyYAML==3.12
prettytable==0.7.2
